+++ src/amp-tags.js
@@ -17,13 +17,33 @@ function showAmpTag({ampStoryPages = true}, pageType, story) {
   return true;
 }
 
+const getDomain = (url) => {
+  try {
+    return new URL(url).origin;
+  }
+  catch (err) {
+    return ""
+  }
+}
+
+/**
+ * StoryAmpTags adds the amphref to stories which support amp.
+ *
+ * @extends Generator
+ * @param {*} seoConfig
+ * @param {(boolean|"public")} seoConfig.ampStoryPages Should amp story pages be shown for all stories (true), not shown (false), or only be shown for public stories ("public"). Default: true
+ * @param {...*} params See {@link Generator} for other Parameters
+ */
 export function StoryAmpTags(seoConfig, config, pageType, data, opts) {
   const story = get(data, ["data", "story"], {});
+  const currentHostUrl = get(data, ["currentHostUrl"], "");
+  // TODO: Remove this condition and always make absolute URL if that's better for AMP discoverability.
+  const ampUrlAppend = seoConfig.appendHostToAmpUrl ? getDomain(currentHostUrl) || config['sketches-host'] : ''
   if(showAmpTag(seoConfig, pageType, story)) {
     return [{
       tag: 'link',
       rel: 'amphtml',
-      href: `/amp/story/${encodeURIComponent(story.slug)}`
+      href: `${ampUrlAppend}/amp/story/${encodeURIComponent(story.slug)}`
     }];
   } else {
     return [];